+++ index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './app.js';
+import App from './ui/app.js';
 
 ReactDOM.render(<App />, document.getElementById('app'));
